Refacto and doc of base_hsm
