+++ crate/hsm/base_hsm/src/error.rs
