+++ crate/hsm/base_hsm/src/hsm_lib.rs
@@ -1,94 +1,49 @@
-use std::{
-    collections::HashMap,
-    ffi::CStr,
-    fmt,
-    fmt::{Display, Formatter},
-    ptr,
-    sync::{Arc, Mutex},
-};
+use std::ptr;
 
 use libloading::Library;
 use pkcs11_sys::*;
 
-use crate::{PError, PResult, SlotManager};
+use crate::{PError, PResult};
 
-struct SlotState {
-    password: Option<String>,
-    slot: Option<Arc<SlotManager>>,
-}
-
-pub struct Utimaco {
-    hsm_lib: Arc<HsmLib>,
-    slots: Mutex<HashMap<usize, SlotState>>,
-}
-
-impl Utimaco {
-    pub fn instantiate<P>(path: P, passwords: HashMap<usize, Option<String>>) -> PResult<Self>
-    where
-        P: AsRef<std::ffi::OsStr>,
-    {
-        let hsm_lib = Arc::new(HsmLib::instantiate(path)?);
-        let mut slots = HashMap::with_capacity(passwords.len());
-        for (k, v) in passwords.iter() {
-            slots.insert(
-                *k,
-                SlotState {
-                    password: v.clone(),
-                    slot: None,
-                },
-            );
-        }
-        Ok(Utimaco {
-            hsm_lib,
-            slots: Mutex::new(slots),
-        })
-    }
-
-    /// Get a slot
-    /// If a slot has already been opened, returns the opened slot.
-    /// To close a slot before re-opening it with another password, call `close_slot()` first
-    pub fn get_slot(&self, slot_id: usize) -> PResult<Arc<SlotManager>> {
-        let mut slots = self.slots.lock().expect("failed to lock slots");
-        // check if we are supposed to use that slot
-        if let Some(slot_state) = slots.get_mut(&slot_id) {
-            if let Some(s) = &slot_state.slot {
-                Ok(s.clone())
-            } else {
-                // instantiate a new slot
-                let manager = Arc::new(SlotManager::instantiate(
-                    self.hsm_lib.clone(),
-                    slot_id,
-                    slot_state.password.clone(),
-                )?);
-                slot_state.slot = Some(manager.clone());
-                Ok(manager)
-            }
-        } else {
-            Err(PError::Default(format!("slot {slot_id} is not accessible")))
-        }
-    }
-
-    pub fn close_slot(&self, slot_id: usize) -> PResult<()> {
-        let mut slots = self.slots.lock().expect("failed to lock slots");
-        slots.remove(&slot_id);
-        Ok(())
-    }
-
-    pub fn get_info(&self) -> PResult<Info> {
-        unsafe {
-            let mut info = CK_INFO::default();
-            let rv =
-                self.hsm_lib.C_GetInfo.ok_or_else(|| {
-                    PError::Default("C_GetInfo not available on library".to_string())
-                })?(&mut info);
-            if rv != CKR_OK {
-                return Err(PError::Default("Failed getting HSM info".to_string()));
-            }
-            Ok(info.into())
-        }
-    }
-}
-
+/// A struct representing a Hardware Security Module (HSM) library interface using PKCS#11.
+///
+/// This struct provides a safe wrapper around the PKCS#11 library functions, managing
+/// the dynamic loading of the HSM library and providing access to cryptographic operations.
+///
+/// # Fields
+///
+/// All fields are PKCS#11 function pointers that correspond to various cryptographic
+/// and key management operations. The fields are marked as `pub(crate)` to allow
+/// access within the crate while maintaining encapsulation.
+///
+/// # Examples
+///
+/// ```no_run
+/// use hsm_common::HsmLib;
+///
+/// let hsm = HsmLib::instantiate("/path/to/hsm/library.so").expect("Failed to load HSM library");
+/// ```
+///
+/// # Safety
+///
+/// This struct handles unsafe FFI calls to the PKCS#11 library internally. The public
+/// interface is designed to be safe to use, but care must be taken when using the
+/// raw function pointers directly.
+///
+/// The library automatically handles initialization and cleanup through the `Drop` trait,
+/// ensuring proper finalization of the HSM when the struct is dropped.
+///
+/// # Methods
+///
+/// - `instantiate<P>`: Creates a new instance of the HSM library
+/// - `initialize`: Initializes the HSM with OS locking capabilities
+/// - `finalize`: Properly closes the HSM connection
+///
+/// # Error Handling
+///
+/// Operations return `PResult<T>`, which is a custom result type for handling
+/// HSM-specific errors. Failed operations typically return `PError` variants
+/// with descriptive error messages.
 #[allow(dead_code)]
 pub struct HsmLib {
     _library: Library,
@@ -130,7 +85,7 @@ pub struct HsmLib {
 }
 
 impl HsmLib {
-    fn instantiate<P>(path: P) -> PResult<Self>
+    pub(crate) fn instantiate<P>(path: P) -> PResult<Self>
     where
         P: AsRef<std::ffi::OsStr>,
     {
@@ -207,52 +162,5 @@ impl HsmLib {
 impl Drop for HsmLib {
     fn drop(&mut self) {
         let _ = self.finalize();
-    }
-}
-
-pub struct Info {
-    pub cryptokiVersion: (u8, u8),
-    pub manufacturerID: String,
-    pub flags: u64,
-    pub libraryDescription: String,
-    pub libraryVersion: (u8, u8),
-}
-
-impl From<CK_INFO> for Info {
-    fn from(info: CK_INFO) -> Self {
-        #[cfg(target_os = "windows")]
-        let flags = u64::from(info.flags);
-        #[cfg(not(target_os = "windows"))]
-        let flags = info.flags;
-        Info {
-            cryptokiVersion: (info.cryptokiVersion.major, info.cryptokiVersion.minor),
-            manufacturerID: CStr::from_bytes_until_nul(&info.manufacturerID)
-                .unwrap_or_default()
-                .to_string_lossy()
-                .to_string(),
-            flags,
-            libraryDescription: CStr::from_bytes_until_nul(&info.libraryDescription)
-                .unwrap_or_default()
-                .to_string_lossy()
-                .to_string(),
-            libraryVersion: (info.libraryVersion.major, info.libraryVersion.minor),
-        }
-    }
-}
-
-impl Display for Info {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Cryptoki Version: {}.{}\nManufacturer ID: {}\nFlags: {}\nLibrary Description: \
-             {}\nLibrary Version: {}.{}",
-            self.cryptokiVersion.0,
-            self.cryptokiVersion.1,
-            self.manufacturerID,
-            self.flags,
-            self.libraryDescription,
-            self.libraryVersion.0,
-            self.libraryVersion.1
-        )
     }
 }
