+++ crate/hsm/base_hsm/src/session/mod.rs
@@ -1,0 +1,5 @@
+mod aes;
+mod rsa;
+
+mod session_impl;
+pub use session_impl::{AesKeySize, HsmEncryptionAlgorithm, RsaKeySize, Session};
