+++ crate/hsm/base_hsm/src/session/rsa.rs
