+++ crate/hsm/base_hsm/src/slots.rs
@@ -11,8 +11,16 @@ use pkcs11_sys::{
 };
 use tracing::warn;
 
-use crate::{utimaco::HsmLib, PError, PResult, Session};
+use crate::{hsm_lib::HsmLib, PError, PResult, Session};
 
+/// A cache structure that maps byte vectors to CK_OBJECT_HANDLE values using an LRU (Least Recently Used) strategy.
+///
+/// This struct wraps a mutex-protected LRU cache that associates object identifiers (as byte vectors)
+/// with their corresponding PKCS#11 object handles. The cache helps improve performance by reducing
+/// repeated lookups of frequently used objects.
+///
+/// The LRU cache automatically removes the least recently accessed entries when it reaches its capacity,
+/// helping to manage memory usage while maintaining quick access to frequently used handles.
 pub struct ObjectHandlesCache(Mutex<LruCache<Vec<u8>, CK_OBJECT_HANDLE>>);
 
 impl Default for ObjectHandlesCache {
@@ -28,6 +36,7 @@ impl ObjectHandlesCache {
         ObjectHandlesCache(Mutex::new(LruCache::new(max)))
     }
 
+    /// Get the object handle for the specified key.
     pub fn get(&self, key: &[u8]) -> Option<CK_OBJECT_HANDLE> {
         self.0
             .lock()
@@ -36,6 +45,7 @@ impl ObjectHandlesCache {
             .copied()
     }
 
+    /// Insert a new object handle into the cache.
     pub fn insert(&self, key: Vec<u8>, value: CK_OBJECT_HANDLE) {
         self.0
             .lock()
@@ -43,6 +53,7 @@ impl ObjectHandlesCache {
             .put(key, value);
     }
 
+    /// Remove an object handle from the cache.
     pub fn remove(&self, key: &[u8]) {
         self.0
             .lock()
@@ -51,6 +62,20 @@ impl ObjectHandlesCache {
     }
 }
 
+/// A manager for a specific PKCS#11 slot in a Hardware Security Module (HSM).
+///
+/// This structure maintains the connection to a specific slot within an HSM,
+/// managing the slot's session and object handles.
+///
+/// # Fields
+///
+/// * `hsm_lib` - A thread-safe reference to the HSM library interface
+/// * `slot_id` - The unique identifier for this HSM slot
+/// * `object_handles_cache` - A thread-safe cache of object handles for this slot
+/// * `_login_session` - An optional authenticated session with the HSM slot
+///
+/// The SlotManager is responsible for coordinating operations on a specific HSM slot,
+/// including session management and object handle caching.
 pub struct SlotManager {
     hsm_lib: Arc<HsmLib>,
     slot_id: usize,
@@ -59,6 +84,21 @@ pub struct SlotManager {
 }
 
 impl SlotManager {
+    /// Create a new SlotManager instance for the specified slot.
+    /// If a login password is provided, the slot will be authenticated with the HSM.
+    /// # Arguments
+    /// * `hsm_lib` - A thread-safe reference to the HSM library interface
+    /// * `slot_id` - The unique identifier for this HSM slot
+    /// * `login_password` - An optional password to authenticate the slot
+    /// # Returns
+    /// * `PResult<SlotManager>` - A result containing the SlotManager instance
+    /// # Errors
+    /// * If the slot cannot be opened or authenticated, an error is returned
+    /// * If the HSM library does not support the necessary functions, an error is returned
+    /// * If the HSM returns an error during session creation or login, an error is returned
+    /// # Safety
+    /// This function calls unsafe FFI functions from the HSM library to open a session and authenticate the slot.
+    /// The function is safe to call, but care must be taken when using the resulting SlotManager instance.
     pub fn instantiate(
         hsm_lib: Arc<HsmLib>,
         slot_id: usize,
@@ -89,6 +129,20 @@ impl SlotManager {
         }
     }
 
+    /// Open a new session with the HSM slot.
+    /// The session can be read-only or read-write, depending on the `read_write` parameter.
+    /// # Arguments
+    /// * `read_write` - A boolean flag indicating whether the session should be read-write
+    /// # Returns
+    /// * `PResult<Session>` - A result containing the new session instance
+    /// # Errors
+    /// * If the session cannot be opened, an error is returned
+    /// * If the HSM library does not support the necessary functions, an error is returned
+    /// * If the HSM returns an error during session creation, an error is returned
+    /// # Safety
+    /// This function calls unsafe FFI functions from the HSM library to open a session.
+    /// The function is safe to call, but care must be taken when using the resulting Session instance.
+    /// The session is automatically closed when the Session instance is dropped.
     pub fn open_session(&self, read_write: bool) -> PResult<Session> {
         Self::open_session_(
             &self.hsm_lib,
