+++ crate/hsm/base_hsm/src/tests.rs
