--- crate/hsm/common/src/session/mod.rs
@@ -1,5 +1,0 @@
-mod aes;
-mod rsa;
-
-mod session_impl;
-pub use session_impl::{AesKeySize, UtimacoEncryptionAlgorithm, RsaKeySize, Session};
